feat: complete refactor pagination query states
