--- apps/app/app/(wallet)/[address]/transactions/(hooks)/usePaginationQueryState.ts
@@ -1,45 +1,0 @@
-// Copyright (C) 2023 Light, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import type { PaginationState } from "@tanstack/react-table";
-import { useQueryState } from "next-usequerystate";
-
-// -----------------------------------------------------------------------------
-// Parser
-// -----------------------------------------------------------------------------
-
-export const paginationParser = {
-  parse(value: string): PaginationState {
-    if (!value) {
-      return {
-        pageIndex: 0,
-        pageSize: 10,
-      };
-    }
-    const [pageIndex, pageSize] = value.split(",").map(Number.parseInt);
-    return { pageIndex, pageSize };
-  },
-  serialize(value: PaginationState): string {
-    return `${value.pageIndex},${value.pageSize}`;
-  },
-};
-
-// -----------------------------------------------------------------------------
-// Hook
-// -----------------------------------------------------------------------------
-
-export const usePaginationQueryState = () => {
-  return useQueryState("pagination", paginationParser);
-};
